refactor: display
- 클래스 구현이 불필요하다고 판단
- Utils로 변경
